requirements.txt
flask
pillow
numpy
reportlab
